Feature in 2 methods; all cleartool stuff in clearcase.ts
